+++ apis/lock_api.py
@@ -14,7 +14,7 @@ LOCK_RETRY_INTERVAL = settings.TASKFLOW_LOCK_RETRY_INTERVAL
 REDIS_URL = settings.TASKFLOW_REDIS_URL
 
 
-logger = logging.getLogger('sodar_taskflow.apis.lock_api')
+logger = logging.getLogger('flask.app')
 
 
 def log_status(lock, unlock=False, failed=False):